requirements.txt
streamlit
pandas
plotly
langgraph
langchain-groq
langchain-core
groq
python-dotenv
requests
numpy
sendgrid
